backend/app/requirements.txt
fastapi
uvicorn[standard]
httpx
qdrant-client
numpy
scikit-learn
python-dotenv
sympy
sqlalchemy
pydantic
aiofiles
typing_extensions
